[upd] - refactor code
