+++ app/Http/Controllers/Dashboard/DashboardNurseController.php
@@ -1,0 +1,13 @@
+<?php
+
+namespace App\Http\Controllers\dashboard;
+
+use App\Http\Controllers\Controller;
+use Illuminate\Http\Request;
+
+class DashboardNurseController extends Controller
+{      
+    public function nurses(){
+        return view('dashboard.nurses');
+    }
+}
