+++ app/Http/Controllers/Dashboard/DashboardUserController.php
@@ -1,28 +1,20 @@
 <?php
 
-namespace App\Http\Controllers;
+namespace App\Http\Controllers\dashboard;
 
-use App\Models\Nurse;
-use App\Models\Transaction;
-use App\Models\User;
+use App\Http\Controllers\Controller;
 use Illuminate\Http\Request;
+use App\Models\User;
 use Illuminate\Support\Facades\Hash;
 use Illuminate\Validation\Rule;
 
-class DashboardController extends Controller
+class DashboardUserController extends Controller
 {
-    public function index() {
-        return view('dashboard.index',[
-            'nurses' => Nurse::with(['availability'])->where('availability_id','=','3')->limit(7)->get(),
-            'transactions' => Transaction::with(['status'])->limit(7)->orderByDesc('id')->get(),
-        ]);
-    }
-
     public function users(){
         return view('dashboard.users');
     }
 
-    public function userDetail(User $user){
+    public function detail(User $user){
         return view('dashboard.user-detail',[
             'user' => $user,
         ]);
@@ -54,13 +46,5 @@ class DashboardController extends Controller
     public function delete(Request $request,User $user){
         User::destroy($user->id);
         return redirect('/dashboard/users')->with('success','User '. $user->name .' has been deleted');
-    }
-
-    public function transactions(){
-        return view('dashboard.transactions');
-    }
-
-    public function nurses(){
-        return view('dashboard.nurses');
     }
 }
